requirements.txt
aiohttp
aiomoex
pandas
matplotlib
tensorflow
finam-trade-api
plotly
yfinance
kaleido
Pillow
chevron